Improve emulator wait for bootcomplete
The problem is we want to be able to know which emulator we just started in order to wait for that one specifically to boot.  There's no real simple way to do this oddly enough.  There used to be a trick where you could pass `-prop emu.uuid=blah` to the emulator start command and then later check each adb device for `adb -s x shell getprop emu.uuid` to find the matching one.  This trick no longer works.  Instead what we're now doing is we know the AVD name we start up so we can just check all the adb devices for the matching one until we find it.  Should be rather ok.