Improve the cmdline-tools ordering
Currently, there are cases with older machines that have not been
updated in a while where the "latest" version is not actually the
latest but an explicit version is.

For example, on the 1ESPT-Ubuntu22.04 CI images, the Android SDK found
is very old and is _just_ the command line tools that were the latest
at the time, which was 9.0.

For building Android apps, we would install the "13.0" tools, but the
"latest" would still be selected. If we tried to update "latest" by
installing it again, then we would get a "latest-2" version.

This PR changes the logic to first try and see if there is a revision
number in the source.properties files and use that instead of the
directory name. This is the more correct mechanism to deterine versions
that just using the directory names. If there is no file or the file is
invalid, then the folder name is used as a fallback.