More correct SDK Download
With the SDK download, we get a cmdline-tools folder but it doesn't contain the package.xml file which seems to be required for the sdk itself to recognize this as an installed package.

If we install it to say `$ANDROID_HOME/cmdline-tools/13.0` and subsequently try to install `cmdline-tools;13.0` it will put them in `$ANDROID_HOME/cmdline-tools/13.0-2` which causes yet more problems.  It also makes it ambiguous which one to try and use.

So, with this change we will now install to `$ANDROID_HOME/cmdline-tools/default`, and then explicitly install the given version of the SDK we downloaded (so the package.xml file exists) and finally, try to delete the original `$ANDROID_HOME/cmdline-tools/default` folder since it's no longer needed.

This also deletes the .zip file when done with it.